+++ tests/Unit/Models/CastMemberTest.php
@@ -25,7 +25,7 @@ class CastMemberTest extends TestCase
 
     public function testCastsAttribute()
     {
-        $casts = ['id' => 'string', 'type' => 'smallint'];
+        $casts = ['id' => 'string', 'type' => 'integer'];
         $castMember = new CastMember();
         $this->assertEquals($casts, $castMember->getCasts());
     }
